change register number and store number to strings, fix filter opened multiple times bug,

+++ app/src/main/java/is/posctrl/posctrl_android/ui/settings/RegisterSelectionFragment.kt
@@ -137,7 +137,7 @@ class RegisterSelectionFragment : BaseFragment() {
             )
             requireContext().showConfirmDialog(confirmText, register.registerNumber.toInt()) {
                 appOptionsViewModel.suspendRegister(
-                        store.storeNumber, register.registerNumber.toInt()
+                        store.storeNumber, register.registerNumber
                 )
                 findNavController().navigateUp()
             }
